the length of a string

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,7 +6,7 @@
  *
  * Return: The length of @str.
  */
-int strlen(const char *str)
+size_t strlen(const char *str)
 {
 
 	size_t length = 0;
